bull_gym/requirements.txt
cryptography
colorama
django
djangorestframework
django-filter
djangorestframework-simplejwt
django-rest-multiple-models
markdown
pymysql
python-dotenv
